Breaking changes to namespace, names, and code organization
