+++ Diwen.Xbrl/Xml/ContextCollection.cs
@@ -19,99 +19,99 @@
 //  You should have received a copy of the GNU Lesser General Public License
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace Diwen.Xbrl
+namespace Diwen.Xbrl.Xml
 {
-	using System;
-	using System.Collections.Generic;
-	using System.Collections.ObjectModel;
-	using System.Globalization;
+    using System;
+    using System.Collections.Generic;
+    using System.Collections.ObjectModel;
+    using System.Globalization;
     using Diwen.Xbrl.Extensions;
 
     public class ContextCollection : KeyedCollection<string, Context>, IEquatable<IList<Context>>
-	{
-		static IFormatProvider ic = CultureInfo.InvariantCulture;
+    {
+        static IFormatProvider ic = CultureInfo.InvariantCulture;
 
-		Instance Instance;
+        Instance Instance;
 
-		public string IdFormat { get; set; }
+        public string IdFormat { get; set; }
 
-		public ContextCollection()
-		{
-			IdFormat = "A{0}";
-		}
+        public ContextCollection()
+        {
+            IdFormat = "A{0}";
+        }
 
-		public ContextCollection(Instance instance)
-			: this()
-		{
-			Instance = instance;
-		}
+        public ContextCollection(Instance instance)
+            : this()
+        {
+            Instance = instance;
+        }
 
-		public new Context Add(Context context)
-		{
-			if (context == null)
-				throw new ArgumentNullException(nameof(context));
+        public new Context Add(Context context)
+        {
+            if (context == null)
+                throw new ArgumentNullException(nameof(context));
 
-			if (context.Entity == null)
-				context.Entity = Instance.Entity;
+            if (context.Entity == null)
+                context.Entity = Instance.Entity;
 
-			if (context.Period == null)
-				context.Period = Instance.Period;
+            if (context.Period == null)
+                context.Period = Instance.Period;
 
-			if (string.IsNullOrEmpty(context.Id))
-			{
-				var exists = false;
-				foreach (var oldContext in this)
-				{
-					if (context.Equals(oldContext))
-					{
-						exists = true;
-						context = oldContext;
-						break;
-					}
-				}
+            if (string.IsNullOrEmpty(context.Id))
+            {
+                var exists = false;
+                foreach (var oldContext in this)
+                {
+                    if (context.Equals(oldContext))
+                    {
+                        exists = true;
+                        context = oldContext;
+                        break;
+                    }
+                }
 
-				if (!exists)
-				{
-					context.Id = NextId();
-					base.Add(context);
-				}
-			}
-			else
-			{
-				base.Add(context);
-			}
+                if (!exists)
+                {
+                    context.Id = NextId();
+                    base.Add(context);
+                }
+            }
+            else
+            {
+                base.Add(context);
+            }
 
-			return context;
-		}
+            return context;
+        }
 
-		public string NextId()
-		{
-			var counter = Count;
-			string id;
-			do
-				id = string.Format(ic, IdFormat, counter++);
-			while (Contains(id));
+        public string NextId()
+        {
+            var counter = Count;
+            string id;
+            do
+                id = string.Format(ic, IdFormat, counter++);
+            while (Contains(id));
 
-			return id;
-		}
+            return id;
+        }
 
-		public void AddRange(IEnumerable<Context> values)
-		{
-			if (values != null)
-				foreach (var item in values)
-					Add(item);
-		}
+        public void AddRange(IEnumerable<Context> values)
+        {
+            if (values != null)
+                foreach (var item in values)
+                    Add(item);
+        }
 
-		protected override string GetKeyForItem(Context item)
-		=> item != null
-				? item.Id
-				: null;
+        protected override string GetKeyForItem(Context item)
+        => item != null
+                ? item.Id
+                : null;
 
-		#region IEquatable implementation
+        #region IEquatable implementation
 
-		public bool Equals(IList<Context> other)
-		=> this.ContentCompare(other);
+        public bool Equals(IList<Context> other)
+        => this.ContentCompare(other);
 
-		#endregion
-	}
+        #endregion
+    }
 }
