+++ Diwen.Xbrl/Xml/FilingIndicator.cs
@@ -19,78 +19,78 @@
 //  You should have received a copy of the GNU Lesser General Public License
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace Diwen.Xbrl
+namespace Diwen.Xbrl.Xml
 {
-	using System;
-	using System.Diagnostics;
-	using System.Xml.Schema;
-	using System.Xml.Serialization;
+    using System;
+    using System.Diagnostics;
+    using System.Xml.Schema;
+    using System.Xml.Serialization;
 
-	[DebuggerDisplay("{Value} : {Filed}")]
-	[Serializable]
-	[XmlRoot("filingIndicator", Namespace = "http://www.eurofiling.info/xbrl/ext/filing-indicators")]
-	public class FilingIndicator : IEquatable<FilingIndicator>
-	{
-		[XmlAttribute("contextRef")]
-		public string ContextRef { get; set; }
+    [DebuggerDisplay("{Value} : {Filed}")]
+    [Serializable]
+    [XmlRoot("filingIndicator", Namespace = "http://www.eurofiling.info/xbrl/ext/filing-indicators")]
+    public class FilingIndicator : IEquatable<FilingIndicator>
+    {
+        [XmlAttribute("contextRef")]
+        public string ContextRef { get; set; }
 
-		[XmlAttribute(AttributeName = "filed", Form = XmlSchemaForm.Qualified,
-			Namespace = "http://www.eurofiling.info/xbrl/ext/filing-indicators")]
-		public bool Filed { get; set; }
+        [XmlAttribute(AttributeName = "filed", Form = XmlSchemaForm.Qualified,
+            Namespace = "http://www.eurofiling.info/xbrl/ext/filing-indicators")]
+        public bool Filed { get; set; }
 
-		[XmlText]
-		public string Value { get; set; }
+        [XmlText]
+        public string Value { get; set; }
 
-		Context contextField;
+        Context contextField;
 
-		[XmlIgnore]
-		public Context Context
-		{
-			get { return contextField; }
-			set
-			{
-				contextField = value;
-				ContextRef = contextField.Id;
-			}
-		}
+        [XmlIgnore]
+        public Context Context
+        {
+            get { return contextField; }
+            set
+            {
+                contextField = value;
+                ContextRef = contextField.Id;
+            }
+        }
 
-		public FilingIndicator()
-		{
-			Filed = true;
-		}
+        public FilingIndicator()
+        {
+            Filed = true;
+        }
 
-		public FilingIndicator(Context context, string value)
-			: this(context, value, true)
-		{
-		}
+        public FilingIndicator(Context context, string value)
+            : this(context, value, true)
+        {
+        }
 
-		public FilingIndicator(Context context, string value, bool filed)
-			: this()
-		{
-			if (context == null)
-				throw new ArgumentNullException(nameof(context));
+        public FilingIndicator(Context context, string value, bool filed)
+            : this()
+        {
+            if (context == null)
+                throw new ArgumentNullException(nameof(context));
 
-			Context = context;
-			Value = value;
-			Filed = filed;
-		}
+            Context = context;
+            Value = value;
+            Filed = filed;
+        }
 
-		public override bool Equals(object obj)
-		=> Equals(obj as FilingIndicator);
+        public override bool Equals(object obj)
+        => Equals(obj as FilingIndicator);
 
-		public override int GetHashCode()
-		=> Value.GetHashCode();
+        public override int GetHashCode()
+        => Value.GetHashCode();
 
-		public override string ToString()
-		=> $"Value={Value}, Filed={Filed}, Context={ContextRef}";
+        public override string ToString()
+        => $"Value={Value}, Filed={Filed}, Context={ContextRef}";
 
-		#region IEquatable implementation
+        #region IEquatable implementation
 
-		public bool Equals(FilingIndicator other)
-		=> other != null
-			&& Filed == other.Filed
-			&& Value.Equals(other.Value, StringComparison.Ordinal);
+        public bool Equals(FilingIndicator other)
+        => other != null
+            && Filed == other.Filed
+            && Value.Equals(other.Value, StringComparison.Ordinal);
 
-		#endregion
-	}
+        #endregion
+    }
 }
