+++ Diwen.Xbrl/Xml/IXbrlObject.cs
@@ -19,12 +19,12 @@
 //  You should have received a copy of the GNU Lesser General Public License
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace Diwen.Xbrl
+namespace Diwen.Xbrl.Xml
 {
 
-	interface IXbrlObject
-	{
-		string Id { get; set; }
-	}
+    interface IXbrlObject
+    {
+        string Id { get; set; }
+    }
 
 }
