+++ Diwen.Xbrl/Xml/SchemaReference.cs
@@ -19,49 +19,49 @@
 //  You should have received a copy of the GNU Lesser General Public License
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace Diwen.Xbrl
+namespace Diwen.Xbrl.Xml
 {
-	using System;
-	using System.Xml.Serialization;
+    using System;
+    using System.Xml.Serialization;
 
-	[Serializable]
-	[XmlRoot(ElementName = "schemaRef", Namespace = "http://www.xbrl.org/2003/linkbase")]
-	public class SchemaReference : IEquatable<SchemaReference>
-	{
-		[XmlAttribute("type", Namespace = "http://www.w3.org/1999/xlink")]
-		public string Type { get; set; }
+    [Serializable]
+    [XmlRoot(ElementName = "schemaRef", Namespace = "http://www.xbrl.org/2003/linkbase")]
+    public class SchemaReference : IEquatable<SchemaReference>
+    {
+        [XmlAttribute("type", Namespace = "http://www.w3.org/1999/xlink")]
+        public string Type { get; set; }
 
-		[XmlAttribute("href", Namespace = "http://www.w3.org/1999/xlink")]
-		public string Value { get; set; }
+        [XmlAttribute("href", Namespace = "http://www.w3.org/1999/xlink")]
+        public string Value { get; set; }
 
-		public SchemaReference()
-		{
-		}
+        public SchemaReference()
+        {
+        }
 
-		public SchemaReference(string type, string value)
-			: this()
-		{
-			Type = type;
-			Value = value;
-		}
+        public SchemaReference(string type, string value)
+            : this()
+        {
+            Type = type;
+            Value = value;
+        }
 
-		public override bool Equals(object obj)
-		=> Equals(obj as SchemaReference);
+        public override bool Equals(object obj)
+        => Equals(obj as SchemaReference);
 
-		public override string ToString()
-		=> $"SchemaReference: Type={Type}, Value={Value}";
+        public override string ToString()
+        => $"SchemaReference: Type={Type}, Value={Value}";
 
-		#region IEquatable implementation
+        #region IEquatable implementation
 
-		public bool Equals(SchemaReference other)
-		=> other != null
-		&& (Type == null ? other.Type == null : other.Type != null && Type.Equals(other.Type, StringComparison.Ordinal))
-		&& (Value == null ? other.Value == null : other.Value != null && Value.Equals(other.Value, StringComparison.Ordinal));
+        public bool Equals(SchemaReference other)
+        => other != null
+        && (Type == null ? other.Type == null : other.Type != null && Type.Equals(other.Type, StringComparison.Ordinal))
+        && (Value == null ? other.Value == null : other.Value != null && Value.Equals(other.Value, StringComparison.Ordinal));
 
-		public override int GetHashCode()
-		=> Value != null ? Value.GetHashCode() : 0;
+        public override int GetHashCode()
+        => Value != null ? Value.GetHashCode() : 0;
 
-		#endregion
+        #endregion
 
-	}
+    }
 }
