+++ Diwen.Xbrl/Xml/Segment.cs
@@ -19,7 +19,7 @@
 //  You should have received a copy of the GNU Lesser General Public License
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace Diwen.Xbrl
+namespace Diwen.Xbrl.Xml
 {
     using System;
     using System.Collections.Generic;
@@ -94,7 +94,7 @@ namespace Diwen.Xbrl
 
         public bool Equals(Segment other)
         => other != null
-        	&& ExplicitMembers.Equals(other.ExplicitMembers)
+            && ExplicitMembers.Equals(other.ExplicitMembers)
             && TypedMembers.Equals(other.TypedMembers);
 
         #endregion
