+++ Diwen.Xbrl/Xml/TypedMember.cs
@@ -19,7 +19,7 @@
 //  You should have received a copy of the GNU Lesser General Public License
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace Diwen.Xbrl
+namespace Diwen.Xbrl.Xml
 {
     using System;
     using System.Xml;
@@ -135,7 +135,7 @@ namespace Diwen.Xbrl
         {
             int result;
             result = string.Compare(Dimension.Name, other.Dimension.Name, StringComparison.Ordinal);
-            
+
             if (result == 0)
                 result = string.Compare(Domain.Name, other.Domain.Name, StringComparison.Ordinal);
