+++ Diwen.Xbrl/Xml/UnitCollection.cs
@@ -19,76 +19,76 @@
 //  You should have received a copy of the GNU Lesser General Public License
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace Diwen.Xbrl
+namespace Diwen.Xbrl.Xml
 {
-	using System;
-	using System.Collections.Generic;
-	using System.Collections.ObjectModel;
-	using System.Linq;
-	using System.Xml;
-	using Diwen.Xbrl.Extensions;
+    using System;
+    using System.Collections.Generic;
+    using System.Collections.ObjectModel;
+    using System.Linq;
+    using System.Xml;
+    using Diwen.Xbrl.Extensions;
 
-	public class UnitCollection : KeyedCollection<string, Unit>, IEquatable<IList<Unit>>
-	{
-		Instance Instance;
+    public class UnitCollection : KeyedCollection<string, Unit>, IEquatable<IList<Unit>>
+    {
+        Instance Instance;
 
-		public UnitCollection() { }
+        public UnitCollection() { }
 
-		public UnitCollection(Instance instance)
-			: this()
-		{
-			Instance = instance;
-		}
+        public UnitCollection(Instance instance)
+            : this()
+        {
+            Instance = instance;
+        }
 
-		public UnitCollection(IEnumerable<Unit> units)
-		{
-			AddRange(units);
-		}
+        public UnitCollection(IEnumerable<Unit> units)
+        {
+            AddRange(units);
+        }
 
-		public void AddRange(IEnumerable<Unit> units)
-		{
-			foreach (var unit in units)
-			{
-				unit.Instance = this.Instance;
-				Add(unit);
-			}
-		}
+        public void AddRange(IEnumerable<Unit> units)
+        {
+            foreach (var unit in units)
+            {
+                unit.Instance = this.Instance;
+                Add(unit);
+            }
+        }
 
-		public void Add(string id, string value)
-		{
-			var idx = value.IndexOf(':');
-			var prefix = value.Substring(0, idx);
-			var ns = Instance.Namespaces.LookupNamespace(prefix);
-			var localname = value.Substring(idx + 1);
-			var measure = new XmlQualifiedName(localname, ns);
-			var unit = new Unit(id, measure);
-			unit.Instance = this.Instance;
-			Add(unit);
-		}
+        public void Add(string id, string value)
+        {
+            var idx = value.IndexOf(':');
+            var prefix = value.Substring(0, idx);
+            var ns = Instance.Namespaces.LookupNamespace(prefix);
+            var localname = value.Substring(idx + 1);
+            var measure = new XmlQualifiedName(localname, ns);
+            var unit = new Unit(id, measure);
+            unit.Instance = this.Instance;
+            Add(unit);
+        }
 
-		public UnitCollection UsedUnits()
-		=> new UnitCollection(this.Where(u => Instance.Facts.Any(f => f.Unit == u)));
+        public UnitCollection UsedUnits()
+        => new UnitCollection(this.Where(u => Instance.Facts.Any(f => f.Unit == u)));
 
-		protected override string GetKeyForItem(Unit item)
-		=> item != null ? item.Id : null;
+        protected override string GetKeyForItem(Unit item)
+        => item != null ? item.Id : null;
 
-		public override bool Equals(object obj)
-		=> Equals(obj as UnitCollection);
+        public override bool Equals(object obj)
+        => Equals(obj as UnitCollection);
 
-		public override int GetHashCode()
-		{
-			int hashCode = 0;
-			foreach (var u in this)
-				hashCode = 31 * hashCode + u.GetHashCode();
+        public override int GetHashCode()
+        {
+            int hashCode = 0;
+            foreach (var u in this)
+                hashCode = 31 * hashCode + u.GetHashCode();
 
-			return hashCode;
-		}
+            return hashCode;
+        }
 
-		#region IEquatable implementation
+        #region IEquatable implementation
 
-		public bool Equals(IList<Unit> other)
-		=> this.ContentCompare(other);
+        public bool Equals(IList<Unit> other)
+        => this.ContentCompare(other);
 
-		#endregion
-	}
+        #endregion
+    }
 }
